fixing warnings, refs #3

+++ src/netedit/elements/network/GNEConnection.cpp
@@ -182,7 +182,7 @@ GNEConnection::getConnectionShapeVertexIndex(Position pos, const bool snapToGrid
         if (shape.front().distanceTo2D(pos) < myNet->getViewNet()->getVisualisationSettings().neteditSizeSettings.connectionGeometryPointRadius) {
             return 0;
         } else if (shape.back().distanceTo2D(pos) < myNet->getViewNet()->getVisualisationSettings().neteditSizeSettings.connectionGeometryPointRadius) {
-            return (shape.size() - 1);
+            return (int)shape.size() - 1;
         } else {
             return -1;
         }
